extract wchar_t definition to rustpython-common
