+++ common/Cargo.toml
@@ -21,3 +21,4 @@ siphasher = "0.3"
 rand = "0.8"
 volatile = "0.3"
 radium = "0.6"
+libc = "0.2.101"
